companion_app/requirements.txt
pyserial
PyQt6
pyqtgraph
numpy
pytest
